requirements.txt
streamlit
llama-cpp-python
pandas